Fixing transition syntax
It could be "from which you have transitioned", or "which you have transitioned to", but not "from which you have transitioned to".  The second option would follow normal syntactic rules, but is not, AFAIK, what is actually intended, since we're dealing with a *previous* state, so it must be either "from which you have transitioned" or "which you have transitioned from".  I'm choosing the former in this pull request since AFAIK there's no way to do a multiple choice request, and I'll spare us both from having a second request, where one or the other is meant to be ignored. :)